+++ src/components/FireMarshal.tsx
@@ -1,6 +1,6 @@
 import {CSSProperties, useContext} from "react";
-import {AppStateContext} from "../domain/appContext";
-import {AppActionsNames, createActionWithPayload} from "../domain/appActions";
+import {AppStateContext} from "../mechanics/appContext";
+import {AppActionsNames, createActionWithPayload} from "../mechanics/appActions";
 import firemanImg from "../images/firemanAndHoseRight.svg"
